+++ CHANGELOG.md
@@ -4,9 +4,9 @@
 
 * token handling
   * load token from environment POMMES_REPOSITORY_TOKEN_<name> variable, not file
-  * github
+  * github + gitlab
     * load github token from "git credentials"
-    * explicitly declare "§§" if it needs a token
+    * explicitly declare "§§" if repository needs a token
 * default repositories
   * the first repository defined in config is the default repository
   * setup automatically defines default repository local = $POMMES_ROOT